eemovel-api/requirements.txt
flask
flask-restx
flask-sqlalchemy
flask-bcrypt
flask-jwt-extended
flask-cors
psycopg2-binary
geoalchemy2
shapely
python-dotenv
pytest
pytest-flask